🎨 moved helper functions to helper file
