+++ src/mqt/predictor/rl/PredictorEnv.py
@@ -12,9 +12,9 @@ from gymnasium import Env
 from gymnasium.spaces import Box, Dict, Discrete
 from pytket.circuit import Qubit
 from pytket.extensions.qiskit import qiskit_to_tk, tk_to_qiskit
-from qiskit import QuantumCircuit, QuantumRegister
-from qiskit.transpiler import CouplingMap, Layout, PassManager, TranspileLayout
-from qiskit.transpiler.passes import ApplyLayout, CheckMap, GatesInBasis
+from qiskit import QuantumCircuit
+from qiskit.transpiler import CouplingMap, PassManager, TranspileLayout
+from qiskit.transpiler.passes import CheckMap, GatesInBasis
 
 from mqt.bench.devices import get_device_by_name
 from mqt.predictor import reward, rl
@@ -246,12 +246,13 @@ class PredictorEnv(Env):  # type: ignore[misc]
                                 pm.property_set["layout"],
                                 pm.property_set["original_qubit_indices"],
                                 pm.property_set["final_layout"],
+                                self.device,
                             )
                     elif action["name"] == "VF2PostLayout":
                         assert pm.property_set["VF2PostLayout_stop_reason"] is not None
                         post_layout = pm.property_set["post_layout"]
                         if post_layout:
-                            altered_qc, pm = self.postprocess_VF2PostLayout(altered_qc, post_layout)
+                            altered_qc, pm = self.postprocess_VF2PostLayout(altered_qc, post_layout, self.layout)
                     else:
                         assert pm.property_set["layout"]
 
@@ -321,33 +322,6 @@ class PredictorEnv(Env):  # type: ignore[misc]
             raise ValueError(error_msg)
 
         return altered_qc
-
-    def postprocess_VF2PostLayout(self, qc: QuantumCircuit, post_layout: Layout) -> tuple[QuantumCircuit, PassManager]:
-        """Postprocesses the given quantum circuit with the post_layout and returns the altered quantum circuit and the respective PassManager."""
-        pm = PassManager(ApplyLayout())
-        assert self.layout is not None
-        pm.property_set["layout"] = self.layout.initial_layout
-        pm.property_set["original_qubit_indices"] = self.layout.input_qubit_mapping
-        pm.property_set["final_layout"] = self.layout.final_layout
-        pm.property_set["post_layout"] = post_layout
-        altered_qc = pm.run(qc)
-        return altered_qc, pm
-
-    def postprocess_VF2Layout(
-        self,
-        qc: QuantumCircuit,
-        initial_layout: TranspileLayout,
-        original_qubit_indices: dict[QuantumRegister, int],
-        final_layout: TranspileLayout,
-    ) -> tuple[QuantumCircuit, PassManager]:
-        """Postprocesses the given quantum circuit with the given layout and returns the altered quantum circuit and the respective PassManager."""
-        postprocessing_action = rl.helper.get_layout_postprocessing_qiskit_pass()(self.device)
-        pm = PassManager(postprocessing_action)
-        pm.property_set["layout"] = initial_layout
-        pm.property_set["original_qubit_indices"] = original_qubit_indices
-        pm.property_set["final_layout"] = final_layout
-        altered_qc = pm.run(qc)
-        return altered_qc, pm
 
     def determine_valid_actions_for_state(self) -> list[int]:
         """Determines and returns the valid actions for the current state."""
